requirements.txt
flask==3.0.0
roboflow==1.1.12
Pillow==9.5.0
numpy==1.24.3
matplotlib==3.8.2